Remove deprecated CLI commands and presets, refactor configuration loading, and streamline server initialization. Consolidate preset management into core functionality for improved maintainability.
