+++ packages/cli/mcpn.config.mjs
@@ -1,0 +1,6 @@
+// MCPN Configuration File
+// Add your configuration options here
+
+export default {
+  "ide": "cursor"
+};
