--- packages/cli/src/utils/presets/thinking.yaml
@@ -1,22 +1,0 @@
-generate_thought:
-  description: "Reflect on a thought and produce a reflection/new set of thoughts"
-  parameters:
-    thought:
-      type: "string"
-      description: "A thought to deeply reflect upon"
-      required: true
-  prompt: |
-    Deeply reflect upon the provided thought step-by-step.
-    Reflect upon the implications/tradeoffs it may have as it relates to my current goals, challenges and our conversation.
-    Do not change anything in our system, just return some thoughts/considerations/analysis based on your reflection of the provided thought.
-
-    Here is the thought to reflect on:
-    {{ thought }}
-
-deep_thinking_mode:
-  description: "Deeply reflect upon the provided idea/problem/question"
-  prompt: |
-    Given the provided conversation, deeply reflect on 6-8 possible tradeoffs, considerations and/or alternative perspectives.
-    For each perspective, produce a comprehensive 2-3 paragraph reflection.
-    At the very end, distill all perspectives into a cohesive and insightful answer
-    Use as much thinking time / inference-time compute as permitted (up to 30 seconds if possible)
