+++ packages/mcpn/src/index.ts
@@ -1,1 +1,4 @@
+// Export server functionality
 export * from "./server";
+export * from "./prompts";
+export * from "./config";
